requirements.txt
yt-dlp
python-telegram-bot
aiohttp
aiofiles
pytest
